Fix: move organization module to main src module
