+++ web/src/modules/app.ts
@@ -50,7 +50,7 @@ import { p2pTransfersReducer } from './user/p2pTransfers';
 import { p2pOrdersReducer } from './user/p2pOrders';
 import { p2pDisputeReducer } from './user/p2pDispute';
 import { feeGroupReducer } from './user/feeGroup';
-import { commercialReducer } from '../custom/modules/user/organization';
+import { commercialReducer } from './user/organization';
 
 export const publicReducer = combineReducers({
     alerts: alertReducer,
