+++ web/src/modules/index.ts
@@ -51,7 +51,7 @@ import { P2POrdersState, rootP2POrdersSaga } from './user/p2pOrders';
 import { P2PDisputeState, rootP2PDisputeSaga } from './user/p2pDispute';
 import { rootFeeGroupSaga, FeeGroupState } from './user/feeGroup';
 import { rootWithdrawLimitsSaga, WithdrawLimitsState } from './public/withdrawLimits';
-import { OrganizationState, rootOrganizationSaga } from 'src/custom/modules/user/organization';
+import { OrganizationState, rootOrganizationSaga } from './user/organization';
 
 export * from './admin/config';
 export * from './admin/markets';
@@ -96,7 +96,7 @@ export * from './user/p2pOffers';
 export * from './user/p2pOrders';
 export * from './user/p2pTransfers';
 export * from './user/p2pDispute';
-export * from '../custom/modules/user/organization';
+export * from './user/organization';
 
 export interface RootState {
     public: {
